backend/enhance: Driver Pool Config for Rental/InterCity
