+++ Backend/app/provider-platform/dynamic-offer-driver-app/Main/src-read-only/Storage/Queries/OrphanInstances/TransporterConfig.hs
@@ -101,6 +101,7 @@ instance FromTType' Beam.TransporterConfig Domain.Types.TransporterConfig.Transp
             fcmConfig = Kernel.External.Notification.FCM.Types.FCMConfig {fcmUrl = fcmUrl', fcmServiceAccount = fcmServiceAccount, fcmTokenKeyPrefix = fcmTokenKeyPrefix},
             freeTrialDays = freeTrialDays,
             includeDriverCurrentlyOnRide = includeDriverCurrentlyOnRide,
+            interCityAdvancedPoolingDurationThreshold = Kernel.Utils.Common.secondsToNominalDiffTime <$> interCityAdvancedPoolingDurationThreshold,
             isAvoidToll = isAvoidToll,
             isPlanMandatory = isPlanMandatory,
             kaptureDisposition = kaptureDisposition,
@@ -246,6 +247,7 @@ instance ToTType' Beam.TransporterConfig Domain.Types.TransporterConfig.Transpor
         Beam.fcmUrl = Kernel.Prelude.showBaseUrl $ (.fcmUrl) fcmConfig,
         Beam.freeTrialDays = freeTrialDays,
         Beam.includeDriverCurrentlyOnRide = includeDriverCurrentlyOnRide,
+        Beam.interCityAdvancedPoolingDurationThreshold = Kernel.Utils.Common.nominalDiffTimeToSeconds <$> interCityAdvancedPoolingDurationThreshold,
         Beam.isAvoidToll = isAvoidToll,
         Beam.isPlanMandatory = isPlanMandatory,
         Beam.kaptureDisposition = kaptureDisposition,
