+++ Backend/app/provider-platform/dynamic-offer-driver-app/Main/src/Storage/CachedQueries/FareProduct.hs
@@ -16,7 +16,7 @@
 module Storage.CachedQueries.FareProduct where
 
 import qualified Domain.Types.Common as DTC
-import qualified Domain.Types.Extra.FareProduct as Domain
+import qualified Domain.Types.Extra.TimeBound as Domain
 import Domain.Types.FareProduct
 import Domain.Types.MerchantOperatingCity (MerchantOperatingCity)
 import qualified Domain.Types.ServiceTierType as DVST
