+++ Backend/dev/migrations-read-only/dynamic-offer-driver-app/transporter_config.sql
@@ -149,3 +149,4 @@ ALTER TABLE atlas_driver_offer_bpp.transporter_config ADD COLUMN use_silent_fcm_
 
 ALTER TABLE atlas_driver_offer_bpp.transporter_config ADD COLUMN fare_recompute_weekly_extra_kms_threshold integer NOT NULL default 20000;
 ALTER TABLE atlas_driver_offer_bpp.transporter_config ADD COLUMN fare_recompute_daily_extra_kms_threshold integer NOT NULL default 5000;
+ALTER TABLE atlas_driver_offer_bpp.transporter_config ADD COLUMN inter_city_advanced_pooling_duration_threshold integer;
